requirements.txt
sounddevice
numpy
scipy
openai
openai-whisper
pyqt5
flask
